requirements.txt
discord.py
langchain
langchain_community
langchain_aws
boto3
python-dotenv